test_requirements.txt
-r rerequirements.txt

pytest==5.4.1
pytest-mock==3.1.0
pytest-asyncio==0.11.0
pytest-aiohttp==0.3.0
